+++ site/content/src/test/scala/exercises/stdlib/EnumerationsSpec.scala
@@ -1,0 +1,11 @@
+package exercises
+
+import stdlib._
+import shapeless.HNil
+
+import org.scalatest.Spec
+import org.scalatest.prop.Checkers
+
+// FIXME: get rid of this if possible
+import org.scalacheck.Shapeless._
+
